bug fixes
Some refactoring and bug fixes
