--- app/src/main/java/com/example/lottikarotti/HighscoreViewHolder.java
@@ -1,8 +1,0 @@
-package com.example.lottikarotti;
-
-import android.view.View;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
