+++ app/src/main/java/com/example/lottikarotti/MainActivity.java
@@ -2,8 +2,6 @@ package com.example.lottikarotti;
 
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,6 +12,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.lottikarotti.Highscore.HighscoreActivity;
 import com.example.lottikarotti.Network.ServerConnection;
 
 import java.net.URISyntaxException;
